backend/requirements.txt
aiohttp==3.5.4
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
googlemaps==4.2.0
idna==2.8
multidict==4.7.5
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.1
requests==2.21.0
sqlparse==0.3.1
urllib3==1.24.3
yarl==1.4.2
